requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
editor==1.6.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inquirer==3.3.0
jiter==0.5.0
openai==1.40.3
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
python-ulid==2.5.0
readchar==4.0.6
regex==2024.7.24
replicate==0.31.0
requests==2.31.0
runs==1.2.2
six==1.16.0
sniffio==1.3.1
tavily-python==0.3.5
tiktoken==0.7.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
xmod==1.8.1
youtube-transcript-api==0.6.2
